feat(be): lưới an toàn — không đầu việc nào âm thầm chết (T128–T146)
Mọi thứ khác trong hệ này đẩy việc đi vì có ai đó **yêu cầu**: người chủ giao, agent nộp,
một lượt rà soát trả về. Còn đúng một hình thù không ai canh, và nó là hình thù làm mất
của dự án cả tuần: *không ai yêu cầu, vì cái sắp yêu cầu đã chết*. Lượt chạy đứt giữa
chừng. Lệnh đánh thức không tới cổng. Máy chủ khởi động lại và mang theo một cái hẹn chỉ
nằm trong bộ nhớ. Đầu việc vẫn ở *đang làm*, bảng vẫn xanh, và **không truy vấn nào trong
lược đồ phân biệt được cái đó với việc đang được làm thật**.

## Động cơ đẩy

Mỗi đầu việc đang chạy mang đúng một động cơ: cái gì sẽ chạm vào nó lần tới, và tới khi
nào (FR-056). Sáu loại, danh sách đóng.

Ba quyết định cấu trúc:

- **Loại nằm trên hàng đầu việc, chi tiết thang leo nằm bảng riêng.** Vòng quét chạy mỗi
  phút, mãi mãi, trên mọi dự án — nên câu hỏi nóng nhất (*động cơ còn hạn không*) phải đọc
  được bằng một lần quét chỉ mục trên một bảng, không nối bảng nào.
- **Tính lại lúc đầu việc đổi trạng thái, không suy trong vòng quét** (QĐ-4). Suy một động
  cơ tốn bốn lượt hỏi sang bốn bảng khác; các ca dùng vốn đã biết chúng vừa hẹn cái gì.
- **Một người ghi duy nhất.** Số vòng chữ ký đã một lần dạy ta cái giá của hai nơi cùng
  ghi một ý nghĩa; không lặp lại.

Có một chỗ tinh: đổi trạng thái và thứ sẽ hành động lên nó **không xảy ra cùng lúc** — đầu
việc vào *chờ làm* trước, lệnh đánh thức được hẹn vài dòng sau, sau khi giao dịch đóng.
Giữa hai khoảnh khắc đó, luật trả lời thành thật rằng "không có gì đẩy". Nên trạng thái để
lại một **động cơ tạm**, và động cơ tạm **luôn có đồng hồ** — kể cả hai loại mà dạng thật
của chúng không có. Đó chính là điểm: đây là phỏng đoán về một việc đang dựng dở, không
phải một lần chờ đã xác minh. Để nó không hết hạn thì "chuyển sang *bị chặn* rồi quên" trở
thành vô hình vĩnh viễn — đúng cái hỏng mà tính năng này sinh ra để đóng lại.

## Vòng quét: hai lượt, cố ý không cân nhau

Lượt rẻ là một truy vấn chỉ mục; trên dự án khoẻ nó trả về rỗng, và đó là ca phổ biến áp
đảo. Lượt thật dựng lại động cơ từ trạng thái sống, **chỉ cho những gì lượt rẻ nêu tên** —
vì bản lưu là một quyết định đã cũ và có thể cũ theo hướng *tốt*: một lượt chạy vừa mở ra.
Nổi chuông chỉ dựa vào bản lưu là chuông báo hão, và chuông báo hão là chuông bị tắt.

Lượt rẻ còn hỏi thêm "đang mang cờ không" — mệnh đề đó cho tiếng chuông một **đường đi
xuống**. Thiếu nó, đầu việc được cứu sẽ có động cơ sống, rơi khỏi hai mệnh đề đầu, và mang
cờ đỏ trên bảng vĩnh viễn.

## Thang phục hồi

Xếp theo giá phải trả và leo từng nấc (FR-059). Luật mang sức nặng là **không nhảy cóc**:
cái thang nhảy thẳng lên người chủ là cái thang hỏi người ta về thứ hệ còn chưa thử lần
nào. Ngân sách Mức 1 đếm theo từng nguyên nhân, giãn dần có trần hai giờ, và tiến triển
thật trả lại **cả** ngân sách chứ không trả từng lần. Lên Mức 3 thì kèm hồ sơ đã thử
(FR-061) — người chủ đọc một lần là trả lời được, thay vì phải đi dựng lại xem hệ đã làm
gì rồi.

## Tuyên treo (FR-062)

Một dòng lượt chạy vẫn ghi RUNNING là lời nói dối thuyết phục nhất trong lược đồ: mọi phép
kiểm khác đều tin nó. Hai ngưỡng chứ không một — *nghi* rồi mới *tuyên* — vì một agent nghĩ
bốn phút giữa hai lệnh là đang làm việc, và một cái máy chém một ngưỡng sẽ giết đúng những
lượt chạy chậm mà khoẻ, sau đó thì không ai để nó bật nữa. Hạn được **đọc lại trong chính
giao dịch ghi**: lượt quét và lượt giết cách nhau một giao dịch, và giết một lượt chạy vừa
sống lại là sai lầm duy nhất vòng này tuyệt đối không được phạm.

Phần đáng giá là phần phục hồi: đóng lượt chạy ma thôi thì đầu việc vẫn nằm đó. Kéo về
*chờ làm* và gọi lại **đúng người phụ trách cũ**, trỏ vào việc kế tiếp đã lưu — nhờ vậy
công việc chạy tiếp từ chỗ dừng chứ không làm lại từ đầu.

## Nhắc ba bậc (FR-065)

Thưa dần, mỗi bậc một lần, và **không bậc nào quyết thay người chủ**. Nhắc mỗi giờ dạy
người đọc rằng hộp thư là tiếng ồn — và một khi họ tin thế thì lần nhắc *đầu tiên* cũng
mất tác dụng. Bậc đã gửi được lưu xuống, nên máy chủ nằm hai ngày rồi bật lại không dội cả
ba lần cùng một giây.

## Dựng lại sau khởi động lại (FR-068)

Chạy **trước** khi vòng quét bắt đầu. Động cơ là lời hứa về tương lai, và tiến trình đã
chết mang theo mọi lời hứa nó từng đưa. Lưới quên mất mình đang canh gì khi tiến trình chết
là **không có lưới nào cả**, vì tiến trình chết chính là sự cố nó sinh ra để đỡ.

## Một lỗi thật lộ ra khi làm

`_unlock_dependents` đang đọc `drive == BLOCKED_BY_TASK` như một dấu hiệu "từng bị chặn" —
lấy một trường suy diễn, có nhớ đệm, làm người gác cho một câu hỏi thật. Đổi cách ghi động
cơ là hai bài kiểm đỏ ngay, và cái đỏ đó đúng: đầu việc mà động cơ ghi thứ khác sẽ **kẹt
mãi mãi** dù mọi việc chặn nó đã xong. Giờ hỏi thẳng bảng phụ thuộc — cạnh nối là sự thật,
động cơ chỉ là bản tóm tắt của nó.

## Bản di trú

Đợt 2 thêm cột `drive` để rỗng, có ghi chú rằng Đợt 6 sẽ lấp. Hôm nay là hôm đó — và một
`drive` rỗng trên đầu việc đang sống chính là thứ vòng quét mới đọc thành *bị rơi*. Không
lấp thì lượt tick đầu tiên sau khi nâng cấp sẽ nổi chuông trên **mọi** đầu việc đang mở của
**mọi** dự án cùng một lúc. Bản lấp cố ý thô, và thô về phía an toàn: nó chỉ nhìn được hàng
đầu việc, nên gán động cơ mà mỗi trạng thái *hàm ý* kèm một hạn ngắn thay vì đoán một hạn
chính xác. FR-058 vẫn được giữ — chỉ là muộn vài phút, đổi lấy một lần nâng cấp không trông
như một sự cố toàn hệ.

548 bài kiểm máy chủ xanh (trước: 536), kiểm kiểu giữ nguyên 159, ruff sạch.

Refs #182

+++ backend/armarius/domain/services/escalation.py
@@ -109,4 +109,4 @@ def backoff_seconds(attempt: int, *, base_seconds: int) -> int:
     """
     if attempt <= 1:
         return base_seconds
-    return min(base_seconds * 2 ** (attempt - 1), MAX_BACKOFF_SECONDS)
+    return int(min(base_seconds * 2 ** (attempt - 1), MAX_BACKOFF_SECONDS))
